duplicated code removed
