+++ frontend/foodbook/src/App.js
@@ -7,7 +7,6 @@ import axios from 'axios';
 import propTypes from 'prop-types';
 
 import Main from 'components/Main';
-import MainTest from 'components/MainTest';
 import FormReview from 'containers/FormReview';
 import ReviewDetail from 'containers/ReviewDetail';
 import Introduce from 'components/Introduce';
@@ -25,8 +24,6 @@ const App = (props) => {
         <Switch>
           <Route path="/introduce" exact component={Introduce} />
           <PrivateRoute path="/main" exact component={Main} history={history} />
-          {/* <Route path="/main" exact component={Main} /> */}
-          <PrivateRoute path="/maintest" exact component={MainTest} history={history} />
           <Route path="/main/upload" exact component={FormReview} />
           <Route path="/main/:id" exact render={(propsIn) => <ReviewDetail store={propsIn.store} history={propsIn.history} match={propsIn.match} />} />
           <Redirect to="/main" />
