+++ frontend/foodbook/src/components/Main.js
@@ -1,132 +1,163 @@
 import React, { Component } from 'react';
 import {
-  Grid,
-  Header,
   Menu,
+  Dropdown,
+  Grid,
+  Container,
 } from 'semantic-ui-react';
-import { NavLink } from 'react-router-dom';
-import propTypes from 'prop-types';
 
-import FriendList from 'containers/FriendList';
-import RealCalendar from 'containers/RealCalendar/RealCalendar';
-import ReviewList from 'containers/ReviewList';
-import FormReview from 'containers/FormReview/FormReview';
-import Logout from 'containers/Logout';
+import Calendar from 'containers/RealCalendar';
+import Category from 'components/Category';
 import Myinfo from 'components/Myinfo';
-import Category from 'components/Category/Category';
+import FormReview from 'containers/FormReview';
+import Feed from './Layouts/Feed';
 
-
-export default class Main extends Component {
+export class Main extends Component {
   constructor(props) {
     super(props);
+
     this.state = {
-      activeItem: 'feed',
+      selectedView: <Feed />,
     };
   }
 
-  handleItemClick = (e, { name }) => this.setState({ activeItem: name })
-
   render() {
-    const { history } = this.props;
-    const { activeItem } = this.state;
+    const handleItemClick = (e, { value }) => {
+      if (value === 'feed') {
+        this.setState({
+          selectedView: <Feed className="feed" />,
+        });
+      } else if (value === 'calendar') {
+        this.setState({
+          selectedView: <Calendar className="calendar" />,
+        });
+      } else if (value === 'category') {
+        this.setState({
+          selectedView: <Category className="category" />,
+        });
+      } /* else {
+        this.setState({
+          selectedView: <Location className="location" />,
+        }); */
+    };
 
-    let mainRenderer = <div />;
+    const friendOptions = [
+      {
+        key: 'Me',
+        text: 'Me',
+        value: 'Me',
+      },
+      {
+        key: 'Jenny Hess',
+        text: 'Jenny Hess',
+        value: 'Jenny Hess',
+      },
+      {
+        key: 'Elliot Fu',
+        text: 'Elliot Fu',
+        value: 'Elliot Fu',
+      },
+      {
+        key: 'Stevie Feliciano',
+        text: 'Stevie Feliciano',
+        value: 'Stevie Feliciano',
+      },
+      {
+        key: 'Christian',
+        text: 'Christian',
+        value: 'Christian',
+      },
+    ];
 
-    switch (activeItem) {
-      case 'feed':
-        mainRenderer = (
-          <div className="main-feed-wrapper">
-            <ReviewList />
-          </div>
-        );
-        break;
+    const viewOptions = [
+      {
+        key: 'feed',
+        text: 'Feed',
+        value: 'feed',
+      },
+      {
+        key: 'location',
+        text: 'Location',
+        value: 'location',
+      },
+      {
+        key: 'calendar',
+        text: 'Calendar',
+        value: 'calendar',
+      },
+      {
+        key: 'category',
+        text: 'Category',
+        value: 'category',
+      },
+    ];
 
-      case 'calendar':
-        mainRenderer = (
-          <div className="main-calendar-wrapper">
-            <RealCalendar />
-          </div>
-        );
-        break;
-
-      case 'location':
-        mainRenderer = (<div className="main-location-wrapper"> location </div>);
-        break;
-
-      case 'type':
-        mainRenderer = (
-          <div className="main-type-wrapper">
-            <Category />
-          </div>
-        );
-        break;
-
-      case 'menu':
-        mainRenderer = (<div className="main-menu-wrapper"> menu </div>);
-        break;
-
-      default: // exceptional Case
-        mainRenderer = (<div className="main-error-wrapper"> error </div>);
-        break;
-    }
+    const { selectedView } = this.state;
 
     return (
       <div className="main">
-        <Grid columns={2} container stackable>
-          <Grid.Row height={10}></Grid.Row>
+        <Menu color="teal" style={{ height: '50px' }}>
+          <Menu.Item> FoodBook </Menu.Item>
+          <Menu.Item>
+                  Change view:
+            <Dropdown
+              scrolling
+              options={viewOptions}
+              placeholder="Change views"
+              defaultValue={viewOptions[0].value}
+              onChange={handleItemClick}
+            />
+          </Menu.Item>
+
+          <Menu.Menu position="right">
+            <Menu.Item>
+                Go to friend&apos;s home:
+              <Dropdown
+                inline
+                options={friendOptions}
+                defaultValue={friendOptions[0].value}
+                style={{ marginLeft: '3px' }}
+              />
+            </Menu.Item>
+            <Menu.Item>
+              <Dropdown
+                scrolling
+                selection
+                search
+                options={friendOptions}
+                placeholder="Add New Friends..."
+                style={{ width: '180px', height: '13px', marginTop: '5px' }}
+                icon="search"
+              />
+            </Menu.Item>
+            <Menu.Item style={{ marginRight: '50px' }}>
+                     Logout
+            </Menu.Item>
+          </Menu.Menu>
+        </Menu>
+        {/*  Top Menu Region */}
+        <Grid>
           <Grid.Row>
-            <Grid.Column width={4}>
-              <Header as="h1"><NavLink to="/main">FoodBook</NavLink></Header>
-              <Logout history={history} />
+            <Grid.Column width={14} style={{ marginLeft: '100px' }}>
+              <Container className="myinfo">
+                <Myinfo />
+              </Container>
+            </Grid.Column>
+            <Grid.Column width={2} style={{ marginLeft: '-800px' }}>
+              <FormReview mode="ADD" fixed={false} />
+
             </Grid.Column>
           </Grid.Row>
-          {/* view select region */}
 
-          <Grid.Row className="wrapper-friend">
-            <Grid.Column width={4}>
-              <Grid.Row style={{ height: '50px' }}></Grid.Row>
-              <Grid.Row>
-                <div className="user-preview">
-                  <div className="ui special cards">
-                    <div className="card">
-                      <div className="content">
-                        <div className="header" style={{ textAlign: 'center' }}> My Info </div>
-                        <br />
-                        <Myinfo />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </Grid.Row>
-              <Grid.Row style={{ height: '50px' }}></Grid.Row>
-              <Grid.Row><FriendList /></Grid.Row>
-            </Grid.Column>
-            {/* Friend Region */}
 
-            <Grid.Column width={9} className="wrapper-reviews">
-              <Grid.Row style={{ height: '50px' }}>
-                <Menu pointing secondary>
-                  <Menu.Item as="a" active={activeItem === 'feed'} name="feed" onClick={this.handleItemClick} id="button-feed">
-                      Feed
-                  </Menu.Item>
-                  <Menu.Item as="a" active={activeItem === 'calendar'} name="calendar" onClick={this.handleItemClick} id="button-calendar">Calendar</Menu.Item>
-                  <Menu.Item as="a" active={activeItem === 'location'} name="location" onClick={this.handleItemClick} id="button-location">Location</Menu.Item>
-                  <Menu.Item as="a" active={activeItem === 'type'} name="type" onClick={this.handleItemClick} id="button-type">Type</Menu.Item>
-                  <Menu.Item as="a" active={activeItem === 'menu'} name="menu" onClick={this.handleItemClick} id="button-menu">Menu</Menu.Item>
-                </Menu>
-              </Grid.Row>
-              <Grid.Row className="add-review">
-                <div className="ui special cards">
-                  <div className="card" style={{ width: '630px' }}>
-                    <div className="content">
-                      <FormReview fixed={false} mode="ADD" />
-                    </div>
-                  </div>
-                </div>
-              </Grid.Row>
-              <br />
-              {mainRenderer}
+          <Grid.Row>
+            <Grid.Column width={2} />
+            <Grid.Column width={12} className="content-wrapper">
+              <center>
+                {selectedView}
+              </center>
             </Grid.Column>
+            <Grid.Column width={2} />
           </Grid.Row>
         </Grid>
       </div>
@@ -134,6 +165,4 @@ export default class Main extends Component {
   }
 }
 
-Main.propTypes = {
-  history: propTypes.objectOf(Object).isRequired,
-};
+export default Main;
